add revalidation logic and role filtering
