+++ knowledge-base-public/src/pages/api/revalidate-sitemap.js
@@ -1,0 +1,18 @@
+// https://nextjs.org/docs/pages/building-your-application/rendering/incremental-static-regeneration
+export default async function handler(req, res) {
+  // Check for secret to confirm this is a valid request
+  if (req.query.secret !== process.env.REVALIDATION_TOKEN) {
+    return res.status(401).json({ message: "Invalid token" });
+  }
+
+  try {
+    // this should be the actual path not a rewritten path
+    // e.g. for "/blog/[slug]" this should be "/blog/post-1"
+    await res.revalidate("/base-de-connaissance/sitemap");
+    return res.json({ revalidated: true });
+  } catch (err) {
+    // If there was an error, Next.js will continue
+    // to show the last successfully generated page
+    return res.status(500).send("Error revalidating");
+  }
+}
